+++ jeap-messaging-infrastructure/src/main/java/ch/admin/bit/jeap/messaging/kafka/signature/publisher/SignatureCertificateHandling.java
@@ -56,8 +56,10 @@ class SignatureCertificateHandling {
     }
 
     private void initMetrics() {
-        validityDaysRemaining = new AtomicLong(certificate.getValidityRemainingDays());
-        signatureMetricsService.initCertificateValidityRemainingDays(() -> validityDaysRemaining.get());
+        if (signatureMetricsService != null) {
+            validityDaysRemaining = new AtomicLong(certificate.getValidityRemainingDays());
+            signatureMetricsService.initCertificateValidityRemainingDays(() -> validityDaysRemaining.get());
+        }
     }
 
     private void initTaskScheduler() {