+++ jeap-messaging-infrastructure/src/test/java/ch/admin/bit/jeap/messaging/kafka/signature/benchmark/SignatureBenchmarkRunner.java
@@ -1,0 +1,9 @@
+package ch.admin.bit.jeap.messaging.kafka.signature.benchmark;
+
+import org.openjdk.jmh.Main;
+
+public class SignatureBenchmarkRunner {
+    public static void main(String[] args) throws Exception {
+        Main.main(args);
+    }
+}