+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.26.0] - 2025-03-06
+
+### Changed
+
+- Add message signing benchmarks
+- Minor bugfixes in message signing
+- Removed conflicting versions of the Amazon Corretto Crypto Provider (ACCP) that could lead to the ACCP being disabled
+
 ## [8.25.0] - 2025-03-06
 
 ### Changed